move render_pass into the swapchain
because:
- it's needed to create framebuffers, and they seem
  like they belong in the swapchain with the image
  views
